requirements.txt
torch(>=1.9.0)
ffmpeg
kaldi-python-io
kaldiio
matplotlib
numpy
pandas
progressbar
progressbar2
pyyaml
scikit-learn
scipy
sklearn
tensorboard
tensorboardx
torchsummary
torchvision
tqdm
setuptools==59.5.0

